Tweak the FLO representation for easier building
This commit alters the `FLO` object format representation in some small
ways to make it easier to build object file output in the hieratika
compiler. It provides access to the context from within the builder for
more in-depth modifications, and also changes the type representation
subtly to make it less complex.