compiler-rt: refactor downcast possibility check to utility function
Bitwise polyfills API uses only u128 as input and output types. The
actual bitwise operations happen on shorter types (e.g. i8 operations
are in fact done on u8 etc). For this reason we make sure that the value
passed in an u128 fits in the target type. This check happens so often,
it is now moved to an utility submodule to reduce code duplication.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>