Add a Cairo -> FlatLoweredObject lifter.
This adds a fairly complete lifter capable of re-representing Cairo
FlatLowered code as FLO objects, which can then be e.g. linked together
with other FLO objects -- including ones taken from Cairo.

There are currently two major limitations on this: a limited number of
cairo interface types are supported (primarily, things like Coupons,
and things that deal with Impls are missing), and this commit lacks
support for lifting Match expressions, which will come in a commit to
follow to avoid introducing any real compiler changes in the main lifter
commit.