Define a custom target for hieratika (#115)
A custom target tells Rustc (and LLVM to an extent) how to compile our
code properly by encoding a bunch of assumptions and behavior
requirements for our target system. In this case, we encode a number of
requirements based on how we plan on using CairoVM and CASM via FLO.

While this commit provides a good starting point, it is extremely likely
that this target will evolve over time.