Test the compiler on `core` and `alloc`
This commit uses the previously-compiled LLVM IR representations of the
rust `core`, `alloc` and `compiler_builtins` libraries to test the
functionality of the hieratika compiler. At the same time it fixes any
bugs that are found through this process.

It also enables the compiler to consistency check the target definition
in the file being compiled against the expected target definition,
ensuring that it does not do unnecessary work on unsupported inputs.

It also implements functionality for working with constants of aggregate
type, thereby closing #91.