Support LLVM 19
This commit bumps our Inkwell dependency to depend on a WIP branch that
adds support for LLVM 19 but without support for any of the new features
that 19 comes with. This means that the version in the nix configuration
is also changed.

This is done in order to support compiling to LLVM IR with the latest
nightlies, which use LLVM 19 and have the most up-to-date `build-std`
support.