compiler-rt: add: fix saturated addition
Overflowing addition did not care about the overflow direction (i.e.
overflow or underflow). It just returned the fact that an overflow
happened or not, as a booloean flag. Meanwhile, a properly implemented
saturated addition must differentiate between an overflow (so it can clamp
to the maximum value) and an underflow (so it can clamp to the minimum
value).

Refactor addition with overflow by returning an enum indicating the
overflow direction. Rewrite the saturated addition to use this information.
Regenerate tests, because they did not consider the direction either,
therefore they were wrong.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>