Fix the handling of global values (#124)
Prior to this commit, global values were being allocated directly as
variables of their type `T`. This is incorrect, as LLVM expects globals
to be of type `ptr` with a `T` able to be loaded from it.

This commit changes the global initialization logic to comply with this
requirement, allocating a pointer of the correct size and then writing
the constant initializer value into that variable. This ensures that
programs expecting to load from globals can now do so correctly.

It also updates the Memory Model documentation to record the solution to
properly handle global pointers in Hieratika.