Implement a prototype name mangler
This commit adds the `hieratika-mangler` crate, responsible for
implementing the hieratika name mangling scheme in a centralized
fashion.

As it exists, the name mangling scheme implemented is _not final_, and
is likely to change and evolve as the requirements are clarified. To
that end, the mangling design docs have not been updated, as no coherent
design process has yet taken place for the mangling scheme.