Design the handling of function pointers (#125)
As we do not have the ability to arbitrarily set the program counter
when compiling via FlatLowered, we have had to come up with a clever
way to emulate calling functions via pointers in Hieratika.

This PR includes that design, which makes certain assumptions that
seem to hold for generic Rust code, but may do not hold for all LLVM
IR. This makes it sufficient for an initial implementation, and for
getting Rust code working, but may later see evolution to better
support more complex notions of function pointers.