Allow `blockaddress` to reference non-locals (#136)
This commit changes the `GetBlockAddressStatement` to allow it to
reference non-local function pointers via name, rather than requiring
the reference target to always be a local `BlockId`.

The previous implementation of the function pointer compilation was not
able to generate relocatable function pointers to functions that are
declared but not defined in the current translation unit. This commit
changes that to now make this case work properly.

This is the first point at which `alloc` is able to be compiled from
LLVM IR to FLO without any crashes or missing features being
encountered.

It also fixes the building of the transient IR input artifacts for
testing the compiler to allow it to work properly on CI. This was not
caught before as there were no tests yet depending on their presence.

The new approach uses a `build.rs` build script to copy them in, and has
also removed one that had accidentally been committed.