Disallow redefining polyfills (#148)
This commit adds a pass that prohibits user code from redefining
polyfills. This is necessary as the actual polyfill names differ from
the opcodes used in LLVM, and hence redefining them needs to be
explicitly disallowed.

This is done using a check pass that runs after the module mapping pass
and uses the module map to check for user-defined functions that clash
with the polyfill names. Note that _declaring_ the polyfills is
explicitly allowed in the case that users want to call them directly.

It also fixes a bug where multiple potential polyfills would end up with
clashing names. It does this by switching the polyfill naming scheme to
instead use the mangler for the types it embeds in the name. This means
we can always generate valid names while also including return type
information for complete disambiguation between polyfills.