Added paste library to deny.toml (#160)
The paste library has been marked as unmaintained, this causes the
cargo deny check to fail. However, this library is currently used by
the crate cairo-lang-runner, therefore it's out of our control. An
exception is the best immediate way forward and a related ticket has
been opened for future tracking of this.