Operate over memory in bytes (#163)
Previously we were not sure whether we would have a bit-addressable or
byte-addressable memory. Now we have made the decision to keep it to the
latter, we want to talk about offsets and sizes in bytes instead of
bits.

To this end, this PR adds implementations of `size_of` and co that
return bytes, in addition to the existing ones in bits, and updates the
compiler to call the byte-oriented methods instead where relevant.