ci: add `scarb check` job
`scarb check` is like `scarb build` but it does not actually build
Sierra from Cairo. Unlike `scarb test`, it analyzes the whole project,
not just test modules, so it can help finding errors `scarb build` would
find but `scarb test` would not.

Since we cannot build the project due to
compiler_rt::crt0::memory::get_allocator function, that is defined but
not yet implemented, this job uses `scarb check`, which is immune to
this problem, rather than `scarb build`, which would fail on the missing
function.

Since now, warnings in the project are not allowed.n the release
profile. This settings makes `scarb check` (but also `scarb build`) fail
(i.e. exit with non-zero code) when warnings are emitted in a release
profile. The dev profile allows warnings, so calling `scarb test` will
not fail if there are unused variables.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>