Bump project dependencies
This commit updates the project's dependencies across all the crates to
the latest versions unless otherwise noted.

This involves a re-write of the components that rely on Chumsky, as the
crate's interface has been re-written for the 0.10.0 release. When
writing the parsing sections, it was incorrect to depend on the 0.9.0
release as this was old and not a reflection of modern parsing practice.
This upgrade fixes this oversight, and also improves parsing speed.