Update the nightly compiler
This commit bumps the version of the nightly compiler used to build the
Rust core libraries. In doing so, a missing feature was discovered in
the hieratika compiler.

To ensure that `core` and `compiler_builtins` continue to compile,
support was implemented for `getelementptr` constant expressions in the
compiler, ensuring that the project is not incompatible with newer
versions of `rustc`.

As part of this work, we clarify the implementation of GEP to allow for
negative offsets in arrays and from base pointers, which is supported by
the spec.