compiler: polyfill: limit bswap to types with even number of bytes
Modify the unary_intrinsic macro to accept also an arbitrary vector of
types. Use that feature to define bswap only over i16/32/48/64/128/256
types, because, as per the LLVM Language reference:

  You can use bswap on any integer type that is an even number of bytes
  (i.e. BitWidth % 16 == 0).

Since bool, i8, i24 and i40 do not conform to this rule, they mangled
names for bswap over these types should not be constructed and added to
the polyfill map. This reduces the overall number of polyfills to 1710.

Remove bswap implementation for i24 and i40 from compiler-rt, because
they will never be used.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>