Lower via Cairo instead of pure CFL
This commit exists as an experiment to see if it is possible to lower
our FLO object format to Cairo source instead of the internal
`FlatLowered` intermediate representation. This is aimed at allowing us
to get something running sooner rather than later, even if it is
fundamentally limited when compared to our initial approach.

It also does some cleanup throughout the codebase and fixes a number of
small bugs that were revealed by the execution of the generated code.

Co-authored-by: Ara Adkins <me@ara.io>