Add a Cairo-compatible mangling scheme
The previous mangling scheme was intended to operate purely in FLO and
Sierra which do not have any care for the way things are named. In the
interests of boxing the work to getting something running, we are
instead hoping to target source Cairo, which requires our naming scheme
to be compatible with Cairo names. This new mangling scheme accounts for
that case.

For the compiler's existing uses of mangling, it now mangles names using
the cairo-compatible scheme.