compiler-rt: introduce state object
A state object of the RTState type keeps the internal state of the
compiler runtime library. As for now, the whole state is just the state
of the memory allocator.

Pass the mutable reference to the object to every polyfill in the
library. For the sake of consistency, every polyfill gets the object,
even if the object is not used.

In unit tests, if the tested polyfill does not use the state object, it
gets an empty object to satisfy the API.

The state object replaces the previous allocator injection mechanism.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>